requirements.txt
appdirs==1.4.3
awesome-slugify==1.6.5
ecdsa==0.13
numpy==1.11.1
packaging==16.8
pandas==0.18.1
paramiko==1.17.2
pycrypto==2.6.1
pyparsing==2.2.0
python-dateutil==2.5.3
pytz==2016.6.1
regex==2016.7.21
six==1.10.0
Unidecode==0.4.19
